Move base.py browser views to browser folder
